move freifunk setup script from /etc/init.d/freifunk_setup to /etc/uci-defaults/50_freifunk_setup
As it turns out, we can write to to the file system at this stage.
Only fastd cannot generate keys at this stage, but there is a more
elegant solution for this included. The only thing we had to give up
is the wifi testing, but we can just expect adhoc and ap to work.